new websocket system implemented for notifications, websocket chat system fixed, update chat and notifications on active implemented, feed filter fixed, loading screen fixed, listeners for event and club implemented - autimatic update via websockets implemented, custom refresh in progress, new notifications implemented, some ui fixes

+++ TogedaTest3/Core/ProfileView/View/ClubsTab.swift
@@ -9,7 +9,7 @@ import SwiftUI
 
 struct ClubsTab: View {
     var userID: String
-    var count: Int32
+    var count: Int64
     
     @Binding var clubs: [Components.Schemas.ClubDto]
